+++ docs/examples/Pythagorean Theorem/index.js
@@ -1,4 +1,7 @@
-figure = new Fig.Figure({ limits: [-4, -3, 8, 6], color: [1, 0, 0, 1] });
+/* globals Fig */
+/* eslint-disable camelcase, no-param-reassign, object-curly-newline */
+
+const figure = new Fig.Figure({ limits: [-4, -3, 8, 6], color: [1, 0, 0, 1] });
 
 
 // ////////////////////////////////////////////////////////////////////////
@@ -9,7 +12,7 @@ figure = new Fig.Figure({ limits: [-4, -3, 8, 6], color: [1, 0, 0, 1] });
 const bc = (content, comment, symbol = null, scale = 0.6, inSize = false) => ({
   bottomComment: {
     content, comment, scale, symbol, inSize,
-  }
+  },
 });
 
 // Add the equation with all it's forms
@@ -41,12 +44,12 @@ figure.add({
       area1: bc('Area_1', 'large square', null, 0.4, true),
       area2: bc('Area_2', 'triangle', null, 0.4, true),
       area3: bc('Area_3', 'small square', null, 0.4, true),
-      A_B: { sup: [{ brac: ['lb', ['A', 'plus_2', 'B'], 'rb'] }, '2_1']},
+      A_B: { sup: [{ brac: ['lb', ['A', 'plus_2', 'B'], 'rb'] }, '2_1'] },
       hAB: [{ frac: ['1_1', 'vinculum', '2_2', 0.7] }, 'A_2', 'B_2'],
       Csq: { sup: ['C', '2_3'] },
       Asq: { sup: ['A_3', '2_4'] },
       Bsq: { sup: ['B_3', '2_6'] },
-      ABExp: ['Asq', 'plus_3', '2_5', 'A_4', 'B_4', 'plus_4', 'Bsq' ],
+      ABExp: ['Asq', 'plus_3', '2_5', 'A_4', 'B_4', 'plus_4', 'Bsq'],
       equals_4t: ['equals', '4', 'times'],
       _4th: ['4', 'times', { frac: ['1_1', 'vinculum', '2_2', 0.7] }],
       s1: { strike: [['2_5', 'A_4', 'B_4'], 'strike1'] },
@@ -54,14 +57,14 @@ figure.add({
     },
     forms: {
       1: ['area1', 'equals_4t', 'area2', 'plus_1', 'area3'],
-      2: [bc('area1', 'A_B', 'brace1'), 'equals_4t','area2', 'plus_1', 'area3'],
-      3: ['A_B', 'equals_4t','area2', 'plus_1', 'area3'],
+      2: [bc('area1', 'A_B', 'brace1'), 'equals_4t', 'area2', 'plus_1', 'area3'],
+      3: ['A_B', 'equals_4t', 'area2', 'plus_1', 'area3'],
       4: ['A_B', 'equals_4t', bc('area2', 'hAB', 'brace1'), 'plus_1', 'area3'],
-      5: ['A_B', 'equals_4t','hAB', 'plus_1', 'area3'],
-      6: ['A_B', 'equals_4t','hAB', 'plus_1', bc('area3', 'Csq', 'brace1')],
-      7: ['A_B', 'equals_4t','hAB', 'plus_1', 'Csq'],
-      8: [bc('A_B', 'ABExp', 'brace1'), 'equals_4t','hAB', 'plus_1', 'Csq'],
-      9: ['ABExp', 'equals_4t','hAB', 'plus_1', 'Csq'],
+      5: ['A_B', 'equals_4t', 'hAB', 'plus_1', 'area3'],
+      6: ['A_B', 'equals_4t', 'hAB', 'plus_1', bc('area3', 'Csq', 'brace1')],
+      7: ['A_B', 'equals_4t', 'hAB', 'plus_1', 'Csq'],
+      8: [bc('A_B', 'ABExp', 'brace1'), 'equals_4t', 'hAB', 'plus_1', 'Csq'],
+      9: ['ABExp', 'equals_4t', 'hAB', 'plus_1', 'Csq'],
       10: ['ABExp', 'equals', bc('_4th', '2_7', 'brace1'), 'A_2', 'B_2', 'plus_1', 'Csq'],
       11: ['ABExp', 'equals', '2_7', 'A_2', 'B_2', 'plus_1', 'Csq'],
       12: ['Asq', 'plus_3', 's1', 'plus_4', 'Bsq', 'equals', 's2', 'plus_1', 'Csq'],
@@ -96,8 +99,18 @@ const makeTriangle = (name, scenario) => ({
       { label: { text: 'A' }, offset: 0.1 },
     ],
     angle: [
-      { curve: { num: 1, sides: 70, radius: 0.25, step: 0.03 }, color: [0.5, 0.8, 0.5, 1] },
-      { curve: { num: 2, sides: 70, radius: 0.25, step: 0.03 }, color: [0.5, 0.5, 1, 1] },
+      {
+        curve: {
+          num: 1, sides: 70, radius: 0.25, step: 0.03,
+        },
+        color: [0.5, 0.8, 0.5, 1],
+      },
+      {
+        curve: {
+          num: 2, sides: 70, radius: 0.25, step: 0.03,
+        },
+        color: [0.5, 0.5, 1, 1],
+      },
       { curve: { autoRightAngle: true, radius: 0.25, step: 0.03 } },
     ],
   },
@@ -123,7 +136,7 @@ const setUpdate = (element) => {
   element.setTransformCallback = () => {
     const rot = element.getRotation();
     element.updateLabels(rot);
-  }
+  };
   element.dim();
 };
 
@@ -178,7 +191,7 @@ figure.elements._start.onClick = () => {
     .opacity({ element: eqn, start: 0.01, target: 1, duration: 1 })
     .dissolveIn({ element: figure.elements._prev, duration: 0.5 })
     .start();
-}
+};
 
 // The prev text tells the user how to proceed. If it is touched, then
 // the equation will go to the previous form
@@ -186,7 +199,7 @@ figure.elements._prev.onClick = () => {
   if (eqn.getCurrentForm().name !== '1') {
     eqn.prevForm({
       duration: 1,
-      ifAnimating: { cancelGoTo: false }
+      ifAnimating: { cancelGoTo: false },
     });
   }
 };
@@ -199,7 +212,7 @@ figure.elements._prev.setTouchable();
 
 // Start by hiding the equation and prev text, and setting all triangles
 // to the same initial position
-eqn.hide()
+eqn.hide();
 figure.elements._prev.hide();
 figure.elements.setScenarios('initial');