+++ docs/examples/Traveling Wave 02 - Sine Waves/tests/fig1/frames.js
@@ -1,0 +1,12 @@
+/* global __title __frames __duration __timeStep */
+__title = 'Example - Traveling Wave 02 - Fig1';
+__duration = -1;
+__timeStep = 0.5;
+__frames = [
+  [0.5, 'touchDown', [1.2, -0.5], 'Touch Down'],
+  [0.5, 'touchMove', [1.2, -0.2], 'Move Arm'],
+  [0.5, 'touchMove', [1.2, 0.2], 'Move Arm'],
+  [0.5, 'touchMove', [1.2, 0.4], 'Move Arm'],
+  [0.5, 'touchMove', [1.2, 0.4], 'Finish Move'],
+  [0.5, 'touchUp'],
+];