+++ docs/examples/Traveling Wave 03 - Velocity Frequency Wavelength/layout.js
@@ -927,6 +927,7 @@ function setupFigure() {
   freqButton1.onClick = () => toggleFrequency(medium1, 1);
   freqButton2.onClick = () => toggleFrequency(medium2, 2);
 
+  figure.elements.hideAll();
   return {
     setMaxTime,
     reset,