+++ docs/examples/Traveling Wave 03 - Velocity Frequency Wavelength/tests/browser.btest.js
@@ -3,6 +3,7 @@ const { tester } = require('../../../../src/tests/browserTester/tester.js');
 tester(
   `${__dirname}/example.html`,
   `${__dirname}/frames.js`,
-  0,
-  0.03,
+  30,
+  0.1,
+  'navigatorFinish',
 );